requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.9.0
click==8.1.3
Deprecated==1.2.14
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Session==0.5.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mysql-connector-python==8.0.33
ordered-set==4.1.0
packaging==23.1
protobuf==3.20.3
Pygments==2.15.1
PyJWT==2.7.0
rich==13.4.2
typing_extensions==4.7.0
Werkzeug==2.3.6
wrapt==1.15.0
